requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.5  
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
bcrypt==4.0.1
stripe== 11.6.0
python-dotenv==0.21.1
psycopg2-binary==2.9.7
Flask-Mail==0.9.1
blinker==1.6.3